# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Quran API Go
  description: |-
    Internal RESTful API that serves Al-Quran data (Arabic text, Indonesian & English translations)
    for the Ilmunara super app.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /surah/{id}/ayah:
    get:
      tags:
        - Ayah
      summary: Get ayahs by surah
      description: Get ayahs from a specific surah with optional range filtering
      operationId: getAyahsBySurah
      parameters:
        - name: id
          in: path
          required: true
          description: Surah ID (1-114)
          schema:
            type: integer
            minimum: 1
            maximum: 114
        - name: from
          in: query
          description: Start ayah number (must use with 'to')
          schema:
            type: integer
            minimum: 1
        - name: to
          in: query
          description: End ayah number (must use with 'from')
          schema:
            type: integer
            minimum: 1
        - name: lang
          in: query
          description: Translation language
          schema:
            type: string
            enum: [id, en]
            default: id
      responses:
        '200':
          description: Surah ayahs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SurahAyahsResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Surah not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /surah/{id}/ayah/{number}:
    get:
      tags:
        - Ayah
      summary: Get ayah by surah and number
      description: Get a specific ayah from a surah by its number within the surah
      operationId: getAyahBySurahAndNumber
      parameters:
        - name: id
          in: path
          required: true
          description: Surah ID (1-114)
          schema:
            type: integer
            minimum: 1
            maximum: 114
        - name: number
          in: path
          required: true
          description: Ayah number within the surah
          schema:
            type: integer
            minimum: 1
        - name: lang
          in: query
          description: Translation language
          schema:
            type: string
            enum: [id, en]
            default: id
      responses:
        '200':
          description: Ayah details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AyahDetailResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Ayah or surah not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /ayah/{id}:
    get:
      tags:
        - Ayah
      summary: Get ayah by global ID
      description: Get a specific ayah by its global ID across all surahs
      operationId: getAyahById
      parameters:
        - name: id
          in: path
          required: true
          description: Global ayah ID
          schema:
            type: integer
            minimum: 1
        - name: lang
          in: query
          description: Translation language
          schema:
            type: string
            enum: [id, en]
            default: id
      responses:
        '200':
          description: Ayah details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AyahDetailResponse'
        '400':
          description: Invalid ayah ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Ayah or surah not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sajda:
    get:
      tags:
        - Ayah
      summary: List sajda ayahs
      description: Get all 15 sajda tilawah ayahs in the Quran
      operationId: listSajdaAyahs
      parameters:
        - name: lang
          in: query
          description: Translation language
          schema:
            type: string
            enum: [id, en]
            default: id
      responses:
        '200':
          description: List of sajda ayahs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/SajdaAyahItem'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /random:
    get:
      tags:
        - Ayah
      summary: Get random ayah
      description: Get a random ayah, optionally filtered by surah
      operationId: getRandomAyah
      parameters:
        - name: surah_id
          in: query
          description: Filter by surah ID (0 = random from all surahs)
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: lang
          in: query
          description: Translation language
          schema:
            type: string
            enum: [id, en]
            default: id
      responses:
        '200':
          description: Random ayah
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AyahDetailResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No ayah found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    SuccessResponse:
      type: object
      required:
        - data
        - timestamp
      properties:
        data:
          type: object
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp in UTC
    ErrorResponse:
      type: object
      required:
        - error
        - code
        - timestamp
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
          enum:
            - bad request
            - not found
            - internal server error
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp in UTC
...
    AyahDetailResponse:
      type: object
      required:
        - id
        - number
        - surah_id
        - number_in_surah
        - text_uthmani
        - translation
        - surah_info
        - juz
      properties:
        id:
          type: integer
          example: 1
        number:
          type: integer
          example: 1
        surah_id:
          type: integer
          minimum: 1
          maximum: 114
          example: 1
        number_in_surah:
          type: integer
          minimum: 1
          example: 1
        text_uthmani:
          type: string
          example: بِسْمِ ٱللَّهِ ٱلرَّحْمَـٰنِ ٱلرَّحِيمِ
        translation:
          type: string
          example: Dengan nama Allah Yang Maha Pengasih, Maha Penyayang.
        surah_info:
          type: object
          required:
            - id
            - name_latin
          properties:
            id:
              type: integer
              example: 1
            name_latin:
              type: string
              example: Al-Fatihah
        juz:
          type: integer
          minimum: 1
          maximum: 30
          example: 1
        sajda:
          type: string
          nullable: true
          description: Sajda type if present
          example: min
        revelation_type:
          type: string
          nullable: true
          example: Meccan
    SurahAyahsResponse:
      type: object
      required:
        - surah
        - ayahs
      properties:
        surah:
          type: object
          required:
            - id
            - number
            - name_latin
          properties:
            id:
              type: integer
              example: 1
            number:
              type: integer
              example: 1
            name_latin:
              type: string
              example: Al-Fatihah
        ayahs:
          type: array
          items:
            type: object
            required:
              - number
              - number_in_surah
              - text_uthmani
              - translation
              - juz
            properties:
              number:
                type: integer
                example: 1
              number_in_surah:
                type: integer
                example: 1
              text_uthmani:
                type: string
                example: بِسْمِ ٱللَّهِ ٱلرَّحْمَـٰنِ ٱلرَّحِيمِ
              translation:
                type: string
                example: Dengan nama Allah Yang Maha Pengasih, Maha Penyayang.
              juz:
                type: integer
                example: 1
              sajda:
                type: string
                nullable: true
                example: min
...
    SajdaAyahItem:
      type: object
      required:
        - id
        - surah_id
        - surah_name
        - number_in_surah
        - text_uthmani
        - translation
        - juz
        - sajda_type
      properties:
        id:
          type: integer
          example: 206
        surah_id:
          type: integer
          example: 7
        surah_name:
          type: string
          example: Al-A'raf
        number_in_surah:
          type: integer
          example: 206
        text_uthmani:
          type: string
          example: إِنَّ ٱلَّذِينَ عِندَ رَبِّكَ
        translation:
          type: string
          example: Sesungguhnya mereka yang ada di sisi Tuhanmu
        juz:
          type: integer
          example: 9
        sajda_type:
          type: string
          description: Type of sajda (obligatory or recommended)
          example: obligatory
...